pyproject.toml
[project]
name = "nonebot-plugin-eitherchoice"
version = "0.2.0"
description = "Let AI help you compare two things."
authors = [{ name = "student_2333", email = "lgc2333@126.com" }]
dependencies = [
    "nonebot2>=2.0.0",
    "nonebot-plugin-htmlrender>=0.2.1",
    "nonebot-plugin-send-anything-anywhere>=0.3.0",
    "pydantic>=1.10.4",
    "httpx[http2]>=0.24.1",
    "beautifulsoup4>=4.12.2",
    "lxml>=4.9.3",
]
requires-python = ">=3.8,<4.0"
readme = "README.md"
license = { text = "MIT" }

[project.urls]
homepage = "https://github.com/lgc-NB2Dev/nonebot-plugin-eitherchoice"

[tool.pdm.build]
includes = []

[tool.pdm.dev-dependencies]
dev = ["nb-cli>=1.0.5", "black>=23.3.0", "ruff>=0.0.260", "isort>=5.12.0"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
